requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
Pillow>=10.0.0
Flask==2.3.3
Flask-Bootstrap4==4.0.2
